refactor(pipeline): convert test imports to use path aliases
Replace deep relative imports (3+ levels) with path aliases (@/pipeline, @/config) in:
- test/unit/pipeline/runner-retry.test.ts
- test/unit/pipeline/stages/rectify.test.ts

Updated barrel exports to support top-level imports:
- Export MAX_STAGE_RETRIES, MAX_STAGE_RESETS from pipeline barrel
- Export _rectifyDeps from stages barrel

Follows project convention: imports 3+ levels deep should use path aliases,
and should target the highest-level barrel when possible.
