+++ test/unit/pipeline/stages/rectify.test.ts
@@ -1,8 +1,8 @@
 // RE-ARCH: keep
 import { describe, expect, test } from "bun:test";
-import { rectifyStage, _rectifyDeps } from "../../../../src/pipeline/stages/rectify";
-import type { PipelineContext } from "../../../../src/pipeline/types";
-import { DEFAULT_CONFIG } from "../../../../src/config";
+import { rectifyStage, _rectifyDeps } from "@/pipeline";
+import type { PipelineContext } from "@/pipeline";
+import { DEFAULT_CONFIG } from "@/config";
 
 function makeCtx(overrides: Partial<PipelineContext> = {}): PipelineContext {
   return {
